完成了设置前驱后续任务的dialog，但是还需要从后台调数据

+++ src/edu/tongji/fiveidiots/ctrl/TaskInfo.java
@@ -147,7 +147,7 @@ public class TaskInfo {
 		this.totaltime = totaltime;
 	}
 
-	public void setNextTaskId(int nextTaskId) {
+	public void setNextTaskId(long nextTaskId) {
 		this.nextTaskId = nextTaskId;
 	}
